+++ src/Channel/Contracts/Channel.php
@@ -14,9 +14,24 @@ declare(strict_types=1);
 
 namespace Vanilo\Channel\Contracts;
 
-interface Channel
+use Vanilo\Contracts\Configurable;
+use Vanilo\Contracts\Merchant;
+
+interface Channel extends Configurable
 {
     public function getName(): string;
 
     public function getSlug(): ?string;
+
+    public function getLanguage(): string;
+
+    public function getCurrency(): ?string;
+
+    public function getDomain(): ?string;
+
+    public function getMerchant(): ?Merchant;
+
+    public function getBillingCountries(): array;
+
+    public function getShippingCountries(): array;
 }