+++ src/Channel/Tests/ChannelTest.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 namespace Vanilo\Channel\Tests;
 
 use Vanilo\Channel\Models\Channel;
+use Vanilo\Contracts\Merchant;
 
 class ChannelTest extends TestCase
 {
@@ -77,5 +78,37 @@ class ChannelTest extends TestCase
         $channel = Channel::create(['name' => 'A bottle of water']);
 
         $this->assertEquals('a-bottle-of-water', $channel->slug);
+    }
+
+    /** @test  */
+    public function it_can_return_the_merchant_from_model_data()
+    {
+        $channel = Channel::create([
+            'name' => 'Ferrari Shop',
+            'email' => 'shop@ferrari.it',
+            'phone' => '+39123456789',
+            'billing_company' => 'Ferrari SpA',
+            'billing_country_id' => 'IT',
+            'billing_city' => 'Maranello',
+            'billing_tax_nr' => 'IT12345678901',
+            'billing_address' => 'Strada Riccovolto 1.',
+            'billing_postalcode' => '41044',
+            'billing_address2' => 'Building F',
+        ]);
+
+        $merchant = $channel->getMerchant();
+        $this->assertInstanceOf(Merchant::class, $merchant);
+        $this->assertEquals('Ferrari SpA', $merchant->getLegalName());
+        $this->assertEquals('IT12345678901', $merchant->getTaxNumber());
+        $this->assertEquals('+39123456789', $merchant->getPhone());
+        $this->assertEquals('shop@ferrari.it', $merchant->getEmail());
+        $this->assertTrue($merchant->isEuRegistered());
+
+        $this->assertEquals('Maranello', $merchant->getAddress()->getCity());
+        $this->assertEquals('Strada Riccovolto 1.', $merchant->getAddress()->getAddress());
+        $this->assertEquals('Building F', $merchant->getAddress()->getAddress2());
+        $this->assertEquals('IT', $merchant->getAddress()->getCountryCode());
+        $this->assertEquals('41044', $merchant->getAddress()->getPostalCode());
+
     }
 }